결제페이지로직수정
